% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_utils.R
\name{constant_matrix}
\alias{constant_matrix}
\title{Creates a matrix filled with a given value.}
\usage{
constant_matrix(val, dimnames, .var.name = varname(val))
}
\arguments{
\item{val}{value for the matrix elements}

\item{...}{matrix dimnames specification}
}
\description{
Creates a matrix filled with a given value.
}
