% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_utils.R
\name{frame2matrix}
\alias{frame2matrix}
\title{Converts data.frame into a matrix
using row_col and col_col as its rows and columns and val_col as its values}
\usage{
frame2matrix(
  df,
  row_col,
  col_col,
  val_col = "w",
  val_default = 0,
  cols = NULL,
  rows = NULL
)
}
\arguments{
\item{df}{data.frame (in a long format) to convert}

\item{row_col}{the Id of the row in the resulting matrix}

\item{col_col}{the Id of the column in the resulting matrix}

\item{val_col}{the values of the matrix elements}

\item{val_default}{the default value if no corresponding row in `df` exist}

\item{cols}{(optional) column ids of the resulting matrix (in a specified order)}

\item{rows}{(optional) row ids of the resulting matrix (in a specified order)}
}
\description{
Converts data.frame into a matrix
using row_col and col_col as its rows and columns and val_col as its values
}
