% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_utils.R
\name{matrix2frame}
\alias{matrix2frame}
\title{Converts a matrix into a 3-column frame (matrix row, matrix column, cell value).}
\usage{
matrix2frame(mtx, row_col = NULL, col_col = NULL, val_col = "w", skip_val = 0)
}
\arguments{
\item{mtx}{matrix to convert}

\item{row_col}{the column of the result containing matrix row names}

\item{col_col}{the column of the result containing matrix column names}

\item{val_col}{the column of the result containing matrix element values}

\item{skip_val}{if not null, the element of the matrix that have this value would not be exported into data frame}
}
\description{
Converts a matrix into a 3-column frame (matrix row, matrix column, cell value).
}
