% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msglm_results.R
\name{msglm.prepare_dims_info}
\alias{msglm.prepare_dims_info}
\title{Generate the dimensions information for the MSGLM report.}
\usage{
msglm.prepare_dims_info(model_data, object_cols = NULL)
}
\arguments{
\item{model_data}{MSGLM model data that is used to get the dimensions info.}

\item{object_cols}{Non-standard object information columns to
include in the reports.}
}
\value{
The named list of data frames with the dimension information.
         The element names are the names of the dimensions,
         Each data frame has `index_<dimname>` key,
         which matches the index along the corresponding dimension in
         the MSGLM model.
}
\description{
Generate the dimensions information for the MSGLM report.
}
\seealso{
process.stan_fit
}
