% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{msglm_data}
\alias{msglm_data}
\title{Title}
\usage{
msglm_data(
  model_def,
  msdata,
  object_ids,
  verbose = model_def$verbose,
  msexperiment_extra_cols = character(0),
  msexperiment_shift_col = "total_shift",
  ...
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
