% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_design.R
\name{msglm_model}
\alias{msglm_model}
\title{Define the experimental design of the MSGLM model}
\usage{
msglm_model(
  conditionXeffect,
  conditions,
  effects,
  msexperimentXeffect = NULL,
  modelobject = c("protgroup", "protregroup", "ptmgroup"),
  quantobject = c("protgroup", "protregroup", "ptmgroup", "pepmodstate"),
  verbose = FALSE
)
}
\arguments{
\item{conditionXeffect}{experimental design matrix,
the rows are conditions, the columns are the effects}

\item{conditions}{the data frame of conditions}

\item{effects}{the data frame of effect with optional prior specification (FIXME)}

\item{msexperimentXeffect}{optional experimental design matrix for
per-MS experiment effect specification. Normally it is not required as `conditionXeffect`
matrix should be enough. It's reserved for the rare cases where the there is a high variability
in biological responses between the biological replicates.}

\item{modelobject}{name of the MS object, which abundance would be modeled by MSGLM}

\item{quantobject}{name of the MS object to take the intensity information from}

\item{verbose}{}
}
\value{
msglm_model object
}
\description{
Define the experimental design of the MSGLM model
}
