% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_utils.R
\name{rbind_all_frames}
\alias{rbind_all_frames}
\title{Binds the rows of the data frames from the list}
\usage{
rbind_all_frames(
  frames_coll_list,
  frame_names = NULL,
  collection_idcol = NULL,
  verbose = FALSE
)
}
\arguments{
\item{frames_coll_list}{list of lists of data frames}

\item{frame_names}{character vector of data frame names (subelements of `frames_coll_list` elements) to bind}

\item{collection_idcol}{name of the additional column that will be set to the name of the `frames_coll_list`
element where the rows are coming from}

\item{verbose}{produce verbose output}
}
\description{
Binds the rows of the data frames from the list
}
