% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_design.R
\name{set_batch_effects}
\alias{set_batch_effects}
\title{Title}
\usage{
set_batch_effects(
  model_def,
  msexperimentXbatchEffect,
  batch_effects = NULL,
  applies_to = c("modelobject", "quantobject"),
  verbose = model_def$verbose
)
}
\arguments{
\item{verbose}{}
}
\value{
updated msglm_model object
}
\description{
Title
}
