% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_design.R
\name{set_contrasts}
\alias{set_contrasts}
\title{Title}
\usage{
set_contrasts(
  model_def,
  metaconditionXcontrast = NULL,
  conditionXmetacondition = NULL,
  contrasts = NULL,
  conditionXcontrast = NULL,
  verbose = model_def$verbose
)
}
\arguments{
\item{verbose}{}
}
\value{
updated msglm_model object
}
\description{
Title
}
