% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_models.R
\name{stan.prepare_data}
\alias{stan.prepare_data}
\title{Coverts the data and experimental design into Stan format.}
\usage{
stan.prepare_data(
  model_data,
  effect_slab_df = 4,
  effect_slab_scale = 2.5,
  obj_labu_min_scale = 1,
  iact_repl_shift_tau = 0.03,
  iact_repl_shift_df = 4,
  hsprior_lambda_a_offset = 0.05,
  hsprior_lambda_t_offset = 0.01,
  batch_effect_sigma = 0.5,
  subbatch_tau = 0.3,
  subbatch_df = 4,
  subbatch_c = 10,
  empty_observation_sigmoid_scale = 1,
  verbose = FALSE
)
}
\arguments{
\item{model_data}{the list with MS data an experimental design}

\item{effect_slab_df}{the *degrees of freedom* for the prior of object effect *slab* regularization parameter}

\item{effect_slab_scale}{the *scale* parameter for the prior of object effect *slab* regularization parameter}

\item{obj_labu_min_scale}{how strongly the estimates below *obj_labu_min* would be penalized (smaller = more stringent)}

\item{iact_repl_shift_tau}{the *tau* parameter for the prior of biochemical log-abundance variation between the replicates}

\item{iact_repl_shift_df}{the *degrees of freedom* parameter for the prior of biochemical log-abundance variation between the replicates}

\item{hsprior_lambda_a_offset}{offset from zero for the "a" parameter of horseshoe priors. Helps NUTS integration step}

\item{hsprior_lambda_t_offset}{offset from zero for the "a" parameter of horseshoe priors. Helps NUTS integration step}

\item{batch_effect_sigma}{sigma parameter for the normal distribution prior of batch effects}

\item{empty_observation_sigmoid_scale}{}

\item{base_input_data}{input data already in Stan format (e.g. MS noise model parameters)}

\item{obj_labu_shift}{the average log-abundance of model objects}

\item{obj_labu_min}{minimal object log-abundance}
}
\description{
Coverts the data and experimental design into Stan format.
}
