% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msglm_results.R
\name{vars_contrast_stats}
\alias{vars_contrast_stats}
\title{Calculate contrasts (linear combination of model effects) using the posterior
MCMC draws.}
\usage{
vars_contrast_stats(
  vars_draws,
  vargroups,
  contrastXvargroup,
  contrasts,
  nsteps = 100L,
  maxBandwidth = NA_real_
)
}
\arguments{
\item{vars_draws}{MCMC draws for the model variables to use for contrast
calculation in `posterior::draws_array` 3D array format.}

\item{vargroups}{grouped data frame that defines how variables are grouped.
The variables are identified by `varspec_ix` column, which corresponds
to the index of the variable in `vars_draws` array.
The existing grouping defines how the variables would be grouped.
The grouping should contain the column having the same name as
the columns dimension name of the `contrastXvalgroup` matrix.
For contrast-specific grouping of the variables the grouping
can optionally include the `contrast` column, and its values should
match the `contrast` column of `contrasts` data frame.}

\item{contrastXvargroup}{the matrix specifying contrasts. The rows are
contrasts (the row names should match the `contrast` column of `contrasts`),
the columns are variable groups (the column names should match the
corresponding column of `vargroups`)}

\item{contrasts}{data frame with all contrasts. The `contrast` is the
obligatory contrast identifier, the obligatory `offset` specifies
what the contrast should be compared with for significance testing.}

\item{nsteps}{(defaults to 100) how many bins to use for the calculation of p-values}

\item{maxBandwidth}{constrain the rule-of-thumb bandwidth for the posterior distribution
if it is above the specified limit}
}
\description{
The procedure allows advanced calculation of contrasts based on contrast-specific
grouping of the variables (provided by `vargroups`). In that case, the
contrasts for all possible combinations of variables in the groups are being
calculated and combined together to form the posterior distribution of the
contrast.
For example, consider the group ``A`` contains variables ``a_1`` and ``a_2``,
and the group ``B`` is ``{b_1, b_2}``. For the ``A - B`` contrast,
the procedure will calculate 4 linear combinations: ``a_1 - b_1``, ``a_1 - b_2``,
``a_2 - b_1`` and ``a_2 - b_2``, then it will combine the resulting 4 sets of
MCMC draws and proceed with calculating the posterior summary.
}
