% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msglm_results.R
\name{vars_pvalues}
\alias{vars_pvalues}
\title{Compare the posterior of each variable in varspecs with its `prior_mean`
and return the p-value for the significance of the difference (w.r.t `tail`)}
\usage{
vars_pvalues(
  vars_draws,
  varspecs,
  tail = c("both", "negative", "positive"),
  nsteps = 100L,
  maxBandwidth = NA_real_
)
}
\arguments{
\item{vars_draws}{MCMC draws from the posterior in `posterior::draws_array` format}

\item{varspecs}{variables to calculate P-values for}

\item{tail}{which comparison to do, one of `both` (the default),
`negative` (``P(x \leq t)``) or `positive` (``P(x \geq t)``),
where ``t`` is the `prior_mean` of the variable.
For `both` the double of the minimal of the two p-values is given.}

\item{nsteps}{(defaults to 100) how many bins to use for the calculation of p-values}

\item{maxBandwidth}{constrain the rule-of-thumb bandwidth for the posterior distribution
if it is above the specified limit}
}
\value{
data frame with the p-value per each varspec
}
\description{
Compare the posterior of each variable in varspecs with its `prior_mean`
and return the p-value for the significance of the difference (w.r.t `tail`)
}
